Move SoliditySign() & SolidityRecover() to evmcompat package
These are now used by the Transfer Gateway code too so doesn't make
sense to keep them in the plasma_cash package.
